refactor(web): lift shared OrgRepos widget into components/ (remove cross-feature page import)
